Complete Place Order Feature

+++ src/main/java/com/bookstore/controller/frontend/shoppingcart/ShoppingCart.java
@@ -36,15 +36,15 @@ public class ShoppingCart {
 		
 		return total;
 	}
-	public double getTotalAmount() {
-		double total = 0.0f;
+	public float getTotalAmount() {
+		float total = 0.0f;
 		
 		Iterator<Book> iterator = cart.keySet().iterator();
 		
 		while(iterator.hasNext()) {
 			Book book = iterator.next();
 			Integer quantity = cart.get(book);
-			double subTotal = quantity* book.getPrice();
+			float subTotal = quantity* book.getPrice();
 			total+=subTotal;
 		}
 		return total;
